tools: refactor build-addons.js to ESM
