+++ doc/contributing/collaborator-guide.md
@@ -232,7 +232,7 @@ There are some other files that touch the build chain. Changes in the following
 files also qualify as affecting the `node` binary:
 
 * `tools/*.py`
-* `tools/build-addons.js`
+* `tools/build-addons.mjs`
 * `*.gyp`
 * `*.gypi`
 * `configure`
